consumers/requirements.txt
kafka-python
minio